requirements/prod.txt
Django==5.1.4
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
django-filter==24.3
Levenshtein==0.26.1
rapidfuzz==3.11.0
boto3==1.35
django-storages==1.14.4
jmespath==1.0.1
python-dateutil==2.9.0.post0
s3transfer==0.11.1
six==1.17.0
urllib3==2.3.0
Pillow==11.1.0
psycopg2==2.9.10
python-dotenv==1.0.1
django-pg-search==0.1.1